refactor : api 분리
